:cool: Bug fixes and polish works done!
1. For preprocessing: Remove the field "TaskName" and "conditions" in
the "Data" field to reduce the size of intermidiate variables, for these
two are not so important. See readsht.m to learn more about this change.
2. For processing: Take into consideration of wrongly recorded trials in
the task DRT, and now removes MRT larger than 2500ms in the subject
level. See the changes in sngstats.m and the newly added function
isnum.m.
3. For plotting: Use a better way to plot error bars for multiple series
data, and now it can accommodate more than 2 series of data, though the
performance might not be preferable.
4. Other minor polishing works: a. Use para instead of curTaskPara; b.
Better logic in processing plot batch work; c. Corresponding remending
works to the binary file taskSetting.xlsx.
:sun_with_face: