:tada: Compatibility of the last change to Prepro.
Since the changes of the Prepro, now Data composes all of its splitRes
into seperate fields (variables) in a table. The compatibility works
have been done for long, now commit them,
And some minor changes to the taskSetting,xlsx are also commited.
Future work might be to transform the data in Excel to json, in order
for better management for the text management is best in Github.